Move API requests for articles
Now they have their own directory :)
