+++ src/app/blog/[slug]/page.ts
@@ -4,7 +4,7 @@ import {
 
 import {
 	getArticleSlugs as generateStaticParams,
-} from '@helpers/requests/get-article-slugs'
+} from '@helpers/requests/articles'
 
 export default Article
 export {generateStaticParams}
