+++ src/components/templates/Home/Home.tsx
@@ -1,6 +1,6 @@
 import {DefaultHead} from '@components/DefaultHead'
 import {PageWrapper} from '@components/PageWrapper'
-import {getIntroArticle} from '@helpers/requests/get-intro-article'
+import {getIntroArticle} from '@helpers/requests/articles'
 import {Addresses} from './Addresses'
 import {ContactForm} from './ContactForm'
 import styles from './Home.module.scss'
